requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-extensions==3.2.3
django-ninja==1.3.0
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
idna==3.10
multidict==6.1.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
